+++ src/nti/externalization/tests/test_representation.py
@@ -8,7 +8,10 @@ from __future__ import division
 from __future__ import print_function
 
 # stdlib imports
+import re
 import unittest
+
+from persistent import Persistent
 
 import fudge
 
@@ -23,6 +26,7 @@ logger = __import__('logging').getLogger(__name__)
 
 # disable: accessing protected members, too many methods
 # pylint: disable=W0212,R0904
+# pylint:disable=attribute-defined-outside-init, useless-object-inheritance
 
 class TestWithRepr(unittest.TestCase):
 
@@ -81,6 +85,70 @@ class TestWithRepr(unittest.TestCase):
         assert_that(r,
                     is_("<nti.externalization.tests.test_representation.Foo("
                         "AttributeError())>"))
+
+    def _normalize_repr(self, r):
+        # Pure-python vs C
+        r = r.replace('nti.externalization.tests.test_representation.', '')
+        # addresses
+        r = re.sub(r'0x[0-9a-fA-F]*', '0xdeadbeef', r)
+        # Python 3.7 removed the trailing , in exception reprs
+        r = r.replace("',)", "')")
+        # Python 2 doesn't have a leading b prefix for byte literals
+        r = r.replace("oid '", "oid b'")
+        return r
+
+    def _normalized_repr(self, o):
+        return self._normalize_repr(repr(o))
+
+    def test_persistent_subclass_default(self):
+        @representation.WithRepr
+        class Foo(Persistent):
+            pass
+
+        o = Foo()
+        r = self._normalized_repr(o)
+
+        assert_that(r,
+                    is_('<Foo object at 0xdeadbeef _p_repr {}>'))
+
+        o._p_oid = b'12345678'
+        r = self._normalized_repr(o)
+
+        assert_that(r,
+                    is_("<Foo object at 0xdeadbeef oid b'12345678' _p_repr {}>"))
+
+        o.a = 1
+
+        r = self._normalized_repr(o)
+
+        assert_that(r,
+                    is_("<Foo object at 0xdeadbeef oid b'12345678' _p_repr {'a': 1}>"))
+
+    def test_persistent_subclass_custom(self):
+        @representation.WithRepr(lambda s: 'Hi')
+        class Foo(Persistent):
+            pass
+
+        o = Foo()
+        r = self._normalized_repr(o)
+        assert_that(r,
+                    is_('<Foo object at 0xdeadbeef _p_repr Hi>'))
+
+    def test_persistent_subclass_raise(self):
+
+        def raise_(self):
+            raise AttributeError()
+
+        @representation.WithRepr(raise_)
+        class Foo(Persistent):
+            pass
+
+        o = Foo()
+        r = self._normalized_repr(o)
+
+        assert_that(r,
+                    is_('<Foo object at 0xdeadbeef _p_repr AttributeError()>'))
+
 
 class TestYaml(unittest.TestCase):